+++ packages/cli/src/commands/process/create.ts
@@ -3,6 +3,7 @@ import {flags} from '@oclif/command'
 import Command from '../../root-command'
 import {findHash} from '../../utils/txevent'
 import {IProcess} from '@mesg/api/lib/process-lcd'
+import {ICoin} from '@mesg/api/lib/transaction'
 
 export default class ProcessCreate extends Command {
   static description = 'Create a process'
@@ -36,6 +37,7 @@ export default class ProcessCreate extends Command {
     if (hashes.length != 1) throw new Error('error while getting the hash of the process created')
     const hash = hashes[0]
     this.spinner.stop(hash)
-    return this.lcd.process.get(hash)
-  }
+    const process = await this.lcd.process.get(hash)
+    return process
+    }
 }