+++ packages/cli/src/commands/process/list.ts
@@ -14,9 +14,15 @@ export default class ProcessList extends Command {
   async run(): Promise<IProcess[]> {
     const {flags} = this.parse(ProcessList)
     const processes = await this.lcd.process.list()
-    cli.table(processes, {
+    const data = await Promise.all(processes.map(async x => ({
+      ...x,
+      account: await this.lcd.account.get(x.address),
+    })))
+    cli.table(data, {
       hash: {header: 'HASH'},
       name: {header: 'NAME'},
+      address: {header: 'ADDRESS'},
+      balance: {header: 'BALANCE', get: (x) => x.account.coins.map(x => `${x.amount}${x.denom}`).join(', ')}
     }, {printLine: this.log, ...flags})
     return processes
   }